set csv io type to double instead of long double

+++ include/io.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <sstream>
 
-std::vector<std::vector<long double>> read_csv(const char* input_file, char delimiter = ',');
-void write_csv(std::vector<std::vector<long double>> values, const char* output_file, char delimiter = ',');
+std::vector<std::vector<double>> read_csv(const char* input_file, char delimiter = ',');
+void write_csv(std::vector<std::vector<double>> values, const char* output_file, char delimiter = ',');
 
 #endif // !IO_H
